LayoutTests/imported/w3c/web-platform-tests/shadow-dom/focus-navigation/tentative/focus-scroller-layout-update-expected.txt

Button

FAIL When checking that element is a scroller, layout information should be up to date. assert_equals: expected Element node <div id="scroller" style="overflow:scroll; width:50px; he... but got Element node <body><button id="button">Button</button>
<div id="scroll...
